+++ tests/Normalizer/FromJson/ContactInfoNormalizerTest.php
@@ -30,7 +30,7 @@ class ContactInfoNormalizerTest extends TestCase
     "bindings": [
       {
         "contactPoint_phoneNumbers": {
-          "value": "+32 9 123 00 02,+32 9 123 00 01,123",
+          "value": "+32 9 123 00 02,+32 9 123 00 01",
           "type": "literal"
         },
         "contactPoint_emailAddresses": {